halo2_gadgets: Move `primitives::*` to `*::primitives`
