+++ halo2_gadgets/src/lib.rs
@@ -15,5 +15,3 @@ pub mod poseidon;
 pub mod sha256;
 pub mod sinsemilla;
 pub mod utilities;
-
-pub mod primitives;
