+++ halo2_gadgets/src/poseidon/primitives/grain.rs
